Add MSVC 2012 and 2013 build settings, make C++11 compatibility changes

+++ ProbeViewer/Source/ChannelViewCanvas/ChannelViewCanvas.hpp
@@ -148,7 +148,7 @@ public:
 private:
     static const Colour backgroundColour;
     
-    size_t drawableSubprocessorIdx;
+    int drawableSubprocessorIdx;
     
     class ProbeViewerCanvas* canvas;
